Woof.  This commit has been coming for a long, ...
Woof.  This commit has been coming for a long, long time -- there have
been long periods of inactivity in MTT development.  :-(

This commit contains much of the "big" work on my to-do items from the
IU MTT meeting:
- implement "reverse" linkage between phases (see samples/odin.ini --
  samples/sample.ini still needs to be updated)
- implement "Trivial" plugins to test MPI "hello world" functionality
  in whatever MPI bindings are provided

There's also a million little fixes and improvements.  See the README
file for the rest of the to-do items that did not yet get done.

This commit was SVN r175.