Add a "variant" field to the test_run table in ...
Add a "variant" field to the test_run table in the new database schema
on the following inspiration: for the past 2 days, I've had failed
infiniband runs but successful TCP runs.  The cause was that I had
borked an IB driver install on my cluster.  So I wouldn't want to
invalidate/delete all my results -- only the IB variants (because the
failed results don't indicate anything about the MPI that was under
test; there was an external factor that was causing it to fail).

"Variant" is a value that the MTT client maintains that has to do with
the expansion of the MPI Details section.  For example, if you have an
OMPI MPI Details section:

  exec = mpirun --mca btl self,&enumerate("tcp", "openib") ...

Then all the TCP runs will be one variant and all the OpenIB runs will
be the other (it actually doesn't matter which is which/which order
the MTT client expands &enumerate() -- looking at the results in the
MTT reporter will make it obvious).  So this schema change means that
the MTT client now needs to submit the variant value along with each
test result.

Adding this field gives a little more fine-grained control to ticket
#7 (user should be able to delete results from the database).

Refs #7.

This commit was SVN r299.