* Use defaults, in case a parameter is not fil...
 * Use defaults, in case a parameter is not filled in by the client.
 * vpath_mode is currently coming over from the client as a string, but
   should eventually be changed to a bool

([http://www.eggheadcafe.com/articles/20050826.asp link to below tip])

== A table should avoid nullable columns ==

Tables can have columns defined to allow null values. A null value
indicates that there is no value. Although it can be useful to allow
null values in isolated cases, it is best to use them sparingly because
they require special handling that increases the complexity of data
operations. For best performance, and to reduce bugs, columns should
ideally be set to NOT NULL. For example, use of the IS NULL keywords in
the WHERE clause makes that portion of the query unable to use an index.
One simple way to avoid nullable columns is to set them as NOT NULL and
provide a default value when creating the column.

This commit was SVN r343.