Fixes #163.
In cases where _prepare_source failed in Build.pm, we blithely ignored
it and proceeded on.  Inevitably, something else would fail later, and
that error would get reported.  But that error was a misnomer, because the
original problem was that _prepare_source failed.

So now we report this better.

This commit was SVN r518.