Refs #170
Added a {{{[Columns]}}} link to the query screen.  (Note:
transitions between phases needs to be ironed out. That is
the new {{{setup_showhide_cols()}}} needs to get worked into
{{{form_carryover()}}}). Required the following be done:

 * Generalized {{{prefs_popup}}} to {{{form_popup}}}.
 * Added applicable fields to
   [source:/trunk/server/sql/views.sql] and
   {{{order_columns()}}}

Maybe it would make more sense for ''all'' the ''Show/Hide''
menu selections in the {{{[Columns]}}} popup? Drawback to
this approach: IE Javascript compatibility (#142).

This commit was SVN r522.