We are currently not reporting on `bitness` for...
We are currently not reporting on `bitness` for Test builds,
so DROP the `bitness` column in `test_build`.  When we do
report on `test_build.bitness`, we should rename both
bitness columns accordingly so they do not have the same
name.  Apparently, we got away with the identical names
because they happened to always be the same value (DEFAULT
of '1'). But now that `get_mpi_install_bitness()` is
working, results were getting hidden (e.g., if
test_build.bitness and mpi_install.bitness were not equal,
{{{SELECT ... FROM test_build_view}}} and {{{SELECT ... FROM
test_run_view}}} would unexpectedly return 0 rows!

This commit was SVN r575.