Fixes #186
At Cisco, our after_each_exec value for the MPI Details section
contains:

{{{
srun -N $SLURM_NNODES killall -9 orted
}}}

This uses SLURM to launch a job on our assigned nodes to ensure to go
cleanup stray orteds, etc.  However, we peridocially saw test run
results that seemed to be killed via -9 for no reason.  That is, the
test seemed to be running ok but then was suddenly killed via -9.
This happened quite randomly, and with low frequency (a dozen times in
40 or 50 thousand runs).

After much investigation, it turned out that the problem was a
combination of factors that effectively delayed the "killall -9 ..."
until ''after'' MTT thought that the after_each_exec step was complete
and, therefore MTT launched the next test.  So the test that was
getting erroneously killed was the ''next'' test -- being killed by
the ''previous'' after_each_exec step.  This is partially MTT's fault
and partially SLURM's fault.

 * MTT dumps the "srun ..." command into a script in /tmp and executes
   it.  A timeout timer of 10 seconds is used to ensure that it
   doesn't hang forever.
 * If the srun is slow to execute for some reason (e.g., a random NFS
   or SLURM delay before the job runs or completes) and MTT's
   after_each_exec timeout expires, MTT will kill the script
   execution.
 * However, srun is stubborn (by design) -- it may sometimes require
   more than one SIGTERM (or whatever) before dying (e.g., because
   it's in a "sensitive" execution location).  So killing the script
   won't necessarily kill the srun -- the srun will stay blocked
   waiting for whatever it was waiting for and eventually execute.
   What we were seeing was that this execution is ''after'' MTT has
   given up on after_each_exec and moved on to the next test.  Hence,
   a false failure.

Note that a side effect of this problem that we saw was that MTT would
kill the script, but then wait several more seconds (complaining each
second) before abondoning the after_each_exec step.  This is because
although the controlling script died, srun itself held open the
stdout/stderr pipes and therefore MTT knew that some child process was
still running.

This is not necessarily a SLURM-specific problem; any after_each_exec
that gets delayed and does not immediately die when MTT kills it may
run into this same problem.  Yuk!

This commit fixes the problem in the following ways:

 * If any of the steps ([before|after]_[each|any]_exec) are a single
   line, MTT will directly launch them (i.e., not create an
   upper-level script and launch that).  Hence, Cisco's INI file now
   contains the following:
{{{
after_each_exec = srun -N &getenv("SLURM_NNODES") ...cleanup script...
}}}
 * Any of the steps can have a "timeout" value in the MPI details
   section.  The default is 10 seconds, but can be anything that
   evaluates to an integer.  Values <=0 result in an infinite
   timeout.  For example:
{{{
after_each_exec_timeout = 30
}}}
 * Any of the steps can have a "pass" value in the MPI details
   section.  The default is:
{{{
<foo>_pass = &and(&cmd_wifexited(), &eq(&cmd_wexitstatus(), 0))
}}}
 * Warnings are issued for steps that do not pass.
 * Per the above bullet, several new funclets were added:
   1. &cmd_wifexited(): Returns 1 if the last command executed exited
      normally, 0 otherwise.
   1. &cmd_wexitstatus(): Returns the argument to exit() (or return
      value from main()) from the last command executed.  Only
      meaningful if &cmd_wifexited() returns 1.
   1. &cmd_wifsignaled(): Returns 1 if the last command executed
      exited via signal.
   1. &cmd_termsig(): Returns the signal number that terminated the
      last command.  Only meaningful if &cmd_wifsignaled() returns 1.

This commit also fixes a few random typos in comments.

This commit was SVN r657.