Fixes #232
Make it so inserting a "!" at the beginning of the search
field translates to using an SQL NOT operator ("!~*"). For a
summary report, when doing a NOT search on a large text
field, an arbitrary digest of the field is shown in the
table (a few leading/trailing characters from the field).

This commit was SVN r685.