Add a "description" field into the MTT, MPI Ins...
Add a "description" field into the MTT, MPI Install, Test Build, and
Test Run sections.  The value in the MTT section is used if it is not
set in the MPI Install/Test Build/Test Run sections (or is empty if it
is not specified at all).

Refs #231

This commit was SVN r849.