This really fixes #287 thanks to Jeff's find.
I ran some basic tests, and we are no longer generating PHP Notice messages.

I also 'fixed' a synchronization problem that arose last night. There is a
race condition between when we check for an existant row and decide to insert
a new entry. The insert will fail in another thread beats us to it, we now
use that as an indicator that we need to search again.
A better way to solve this is to involve transactions, but for now this should
fix the problem.

This commit was SVN r958.