Fixes #313 (still need to commit to coll bakeof...
Fixes #313 (still need to commit to coll bakeoff branch).  Add a new
funclet &rm_name() that will return the name of the ''resource
manager'' -- it will never return "hostfile" or "hostlist" (because
these are not resource managers).  In the case of hostfile/hostlist,
&rm_name() will return "none".  Otherwise, it will return exactly the
same thing as &env_name().

I then changed the default value of the resource_manager field in the
MPI Defaults section to be &rm_name() (instead of &env_name(), which
is was before).

This commit was SVN r1038.