Consolidate some extra server-level debugging w...
Consolidate some extra server-level debugging with a new module
parameter: mttdatabase_server_debug.  Some extra --verbose messages
will be printed during the reporter phase (to the MTTDatabase) if set
to true.  If not set to a true value, these messages are only output
if --debug is used.

This commit was SVN r1114.