Print the total time spent executing SQL querie...
Print the total time spent executing SQL queries. This allows us to tell if
it is the HTML/PHP that is adding to the time, or the execution of the SQL
query by the backend database.

I was noticing that over a slow network connection the Total execution time
was quite large, even though the SQL exeuciton time remained small.

This commit was SVN r1155.