Add new funclet &mpi_details() that takes a str...
Add new funclet &mpi_details() that takes a string argument.  The
string corresponds to a field name that was set in the Test run
section.  It is best explained through example:

{{{
[MPI Details: foo]
exec = mpirun &mpi_details("mpirun_argv") ...

[Test run: bar]
mpi_details:mpirun_argv = --bynode
}}}

The "mpi_details:" prefix to fields in the test run section tells MTT
that the string key after the : is the parameter passed to
&mpi_details() that is will be used to retrieve the corresponding
value in the MPI details section.  In the above case, the "bar" test
run section was passing "--bynode" to the mpirun command through the
"mpirun_argv" key.

This mechanism replaces the less flexible "alloc" mechanism that
allowed restricted passing of some values between test run and mpi
details sections.

This commit was SVN r1168.