Add a new INI field: ompi_snapshot_version_file...
Add a new INI field: ompi_snapshot_version_file.  If this field is
defined, it's a filename where the MTT client can store the single
last version of tarball that was obtained in that MPI Get phase.  This
location should be outside the scratch tree (i.e., durable if the
scratch tree is removed or otherwise unavailable for the next MTT
run).

This file is checked before downloading a new tarball; if the version
that we would download is the same as the version in the cached file,
we rule that there is no new version and this MPI Get phase is
trimmed.

This is most useful for those of us who are MTT testing OMPI with
different scratch trees every night.  If no developer commits anything
to the OMPI tree during the next day, no new tarball will be
generated.  Before this field, MTT would happily test the same tarball
that it tested last night.  Now, with this field, MTT will know that
it's already been tested and skip it.

Go green!  :-)

This commit was SVN r1283.