This patch adds a new boolean flag "email_detai...
This patch adds a new boolean flag "email_detailed_reports" which sends out the detailed reports of each phase as attachments.
It pulls in two files from CPAN to encode the email as a multipart (MIME) message.
The *.txt files generated as the output of each stage are attached by the email reporter to the summary email.

Signed-off-by: Abhishek Kulkarni <adkulkar@osl.iu.edu>

This commit was SVN r1386.