Ok, I think I broke things with r1591.  This co...
Ok, I think I broke things with r1591.  This commit should fix things.

The issue was that Test::Run::Run was getting called ''twice'' (and
displaying its output twice), which was very, very confusing.  It also
fired up all the infrastructure of Test::Run::Run (e.g., calling
funclets), which further added to the confusion.

It looks like the goal in doing this was to increment
$MTT::Globals::Internals->{total_tests_counter} with the actual number
of variants run.  This commit does that without invoking
Test::Run::Run twice.

Hopefully, I've understood the intent of the original work properly;
perhaps miked can confirm that this commit still has the intended
end-effect as r1433.

This commit was SVN r1594.

The following SVN revision numbers were found above:
  r1433 --> db1f1a0ab9dbb24b7edb1fc5a71a682668127735
  r1591 --> d61a2b6ce1971a24b4b032da4e68cacf0d64f1db