Diskfree.pm: fix df handle leak
This left a zombied "df" process in the process table.  The Perl
runtime eventually cleaned it up, but it was weird to see a zombied df
in ps listings.

There's no point in reporting this bug upstream; according to the
Changes file, v0.06 was published in 1998.  There's been no activity
on http://search.cpan.org/dist/Filesys-DiskFree/DiskFree.pm since
v0.06.