Analyze: allow treating timeouts as failures
Per request from the Open MPI community, add a parameter
"treat_timeouts_as_fail" to be set in Test Run sections that, upon
result of a timeout, will override that result and classify it as a
failure.

The rationale is that timeouts are not frequently examined in the web
reporter, and at least some kinds of test should treat timeouts as
failures.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>