Commit to ensure we can pass multiple SLURM options
The same problem found in ALPS plugin is also in SLURM plugin. the "options" parameter was being passed in as a whole string rather than being broken apart into multiple arguments. This was fixed by splitting the options string into multiple strings before appending the options to the list of arguments.