Any delimiter (not just ":") can now be used in INI files
Code before required ":" to be used as a delimiter between stage and name. Based on how the information was stored in the log, the delimiter didn't make sense. Also, the title (title=stage+name) with ":" in it was causing an issue with some makefiles. This fix provides freedom in title naming.

[MTTDefaultsIsAVeryCoolStage] is now a valid title.

[TestRun--projectAwesome] is also now a valid title.

As well as [TestGet:GrabMyFiles]

Signed-off-by: Richard Barella <richard.t.barella@intel.com>