Refactor plugin loading to use pathlib
Fixes #435:
This fixes an issue where relative plugin directories would not load
correctly. Furthermore, it simplifies the flow of the program
significantly, making it much more readable.
The changes to TestDef.py also fix an issue where the error reported was
not always the error which occurred. In addition to catching
non-existent directories, it caught any error which occurred in
LoadClasses.load() and gave unhelpful responses.
Per a conversation with @rhc54, plugins in the given plugin
directories will be loaded, as opposed to just their subdirectories.

pathlib and python-hostlist were added to pyenv.txt. pathlib was included as
it is not standard in python <= 3.4. python-hostlist was an existing
requirement which was not in pyenv.txt.

Signed-off-by: Peter Gottesman <pgottesm@cisco.com>