Fixes #798 moved fh var from global to instance
fh (file handle) variable is defined in the __init__ method and as such
only one variable is createdi no matter how many instances of TextFile
are created.

Multiple calls to the TextFile plugin will set this variable to either
stdout or an open file handle, the latter one replacing the previous.

Moved the variable defintion into the execute method, this will allow for
one fh variable to be created per instance of TextFile.

Signed-off-by: Bill Weide <william.c.weide@intel.com>