Trigger travis build
This commit is to trigger the travis build and deployment of the doxygen
documentation. It is in combination with the push of gh-pages with the
html and latex delected to clear out the case sensitive file names. The
gh-pages commit needs to be merged with gh-pages before this commit is
merged with master. Hopefully

Fixes #714

Signed-off-by: Deb Rezanka <rezanka@lanl.gov>