Add support for passing envars to make
The Autotools plugin should only execute configure if the package
actually contains a configure script - otherwise, skip that step. Add
support for setting envars for the "make" step - e.g., to specify the
wrapper compiler to use.

Update the DefaultTestBuild plugin to accept "make_envars" so it can
pass them to the Autotools plugin

Signed-off-by: Ralph Castain <rhc@pmix.org>