Have PRRTE record its name and the middleware version
Use the dependencies to identify the middleware being used to support
the tests being executed - use that to define the version field. Use the
PRRTE build informationto set the name field. Thus, the reporter will
show the PRRTE version in the "MPI name" field and the middleware
version in the "MPI version" field

Signed-off-by: Ralph Castain <rhc@pmix.org>