Use a different method for timeout detection
Apparently, Python 2 does not have the subprocess "timeout" feature in
it. One solution would be to have everyone install the subprocess32
module that backported such features to Python 2, but that would be
intrusive. So instead use a threaded model for tracking timeout.

Little more cumbersome and involves another thread, but appears to be
working just fine under test.

Signed-off-by: Ralph Castain <rhc@pmix.org>