Mark Git repos with the hash of their checkout
When checking out a Git repo, record the hash of the head of the branch
or PR we are sitting on. Use that value to "tag" the dependencies in the
PRRTE and PMIxUnit plugins so they can be used to replicate the tests at
a later time.

Fix a bug in the IUDatabase reporter so it correctly reports the "np"
field from a test.

Signed-off-by: Ralph Castain <rhc@pmix.org>