Correctly set the return result so the IUDatabase reporter can report it
The IUDatabase reporter needs slightly different information to
correctly set the result codes in the MTT database. You need to set the
'results' entry in the log to one of the predefined values. I have tried
to hit the right spots in these two commonly used plugins, but someone
will need to validate that I got them and fill in any that are missing.

Fixes #845

Signed-off-by: Ralph Castain <rhc@pmix.org>