Be more flexible in how we take input
Accept lists of tests expected to fail and tests to skip that are
delimited by commas, spaces, or tabs. Instead of having a separate list
of expected return codes for failed tests, specify those in the same
place where we list those tests:

fail_tests = ftest1:3 ftest2 ftest3:-5

where we specify the expected return code with a ":code" suffix to the
name of the test. Tests without an expected return code will be marked
as "passed" if they fail regardless of what they return.

Add a "dryrun" option to the IUDatabase reporter so we can see what it
would have sent without actually sending anything to the db.

Signed-off-by: Ralph Castain <rhc@pmix.org>