Server code cleanup
mtt/server files edited for compatability with python 2 and 3. Unused gds directory was deleted. It can be found in mtt-legacy if needed. Unused partitions in the database were commented out in mtt/server/sql/support/create-partitions-test-run.pl. Documentation is still a work in progress but some updates have been made.

Signed-off-by: Deb Rezanka <rezanka@lanl.gov>