add support for checkpoint/restart of mtt logfiles
This feature adds support for checkpointing MTT logfiles as part of a MTT workflow
wherein some stages of an MTT workflow must be done on one set of system resources -
in particular login nodes to typical HPC DoE systems which have internet access -
but for other stages, such as running tests, MTT has to be run on backend nodes
with no internet access.

Pickle is used to serialize/deserialize the logfile irrespective of its structure.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>