Added check for slurm job step abort time out
This covers a situation where the job is aborted but the slurm cleanup times out, leaving
the job status as 0, here is the typical output

     stderr: srun: Force Terminated job 6050
     stderr: srun: Job step aborted: Waiting up to 32 seconds for job step to finish.
     stderr: srun: error: Timed out waiting for job step to complete

A dict was added to the slurm SLURM class to contain an error string to search for and
results and test return codes.  This is passed to the LauncherMTTTool, runTests function
where this additional check can be made.  The code is structured this way to allow the
SLURM plugin to do these checks and allow other launcher plugins to not do these checks.

Signed-off-by: Bill Weide <william.c.weide@intel.com>