Clone a separate repo for each TestGet section
This change alters the git clone behavior for TestGet sections.

Currently, if TestGet:XXX and TestGet:YYY section requires the same git repository,
the repo is only cloned once, which is unlikely the desired behavior.

With this change, each TestGet section will have a separate clone of the repo. This
prevents dependent TestRun conflicts.